ai_service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pymupdf==1.23.5
python-docx==0.8.11
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.25.2
nltk==3.8.1
spacy==3.7.2
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
flask-testing==0.8.1
